make those angles double to avoid warnings and potentially get better results too

+++ source/xatlas/xatlas.cpp
@@ -2715,9 +2715,9 @@ public:
 
 		// use the corner whose angle is the closest to a right angle,
 		// as that will give the most stable results for the cross product calculation
-		const float diff0 = abs(M_PI_2 - angle(e00, e01));
-		const float diff1 = abs(M_PI_2 - angle(e10, e11));
-		const float diff2 = abs(M_PI_2 - angle(e20, e21));
+		const double diff0 = abs(M_PI_2 - angle(e00, e01));
+		const double diff1 = abs(M_PI_2 - angle(e10, e11));
+		const double diff2 = abs(M_PI_2 - angle(e20, e21));
 		Vector3 normalAreaScaled;
 		if (diff0 <= diff1 && diff0 <= diff2) {
 			normalAreaScaled = cross(e00, e01);
